LPS-133951 move code on new-js to js
