+++ modules/apps/data-engine/data-engine-taglib/src/main/java/com/liferay/data/engine/taglib/servlet/taglib/DataLayoutBuilderTag.java
@@ -158,7 +158,7 @@ public class DataLayoutBuilderTag extends BaseDataLayoutBuilderTag {
 
 	private String _getModule() {
 		if (Validator.isBlank(getModule())) {
-			return "data_layout_builder/new-js/App";
+			return "data_layout_builder/js/App";
 		}
 
 		return getModule();
@@ -174,7 +174,7 @@ public class DataLayoutBuilderTag extends BaseDataLayoutBuilderTag {
 
 	private String _getPluginEntryPoint(String value) {
 		return DataLayoutTaglibUtil.resolveModule(
-			"data-engine-taglib/data_layout_builder/new-js/plugins/" + value +
+			"data-engine-taglib/data_layout_builder/js/plugins/" + value +
 				"/index");
 	}
