+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/js/plugins/fields-sidebar/components/FieldsSidebarBody.js
@@ -15,9 +15,9 @@
 import {useConfig} from 'data-engine-js-components-web';
 import React from 'react';
 
-import Sidebar from '../../../../js/components/sidebar/Sidebar.es';
 import FieldSetList from '../../../components/field-sets/FieldSetList';
 import FieldTypeList from '../../../components/field-types/FieldTypeList.es';
+import Sidebar from '../../../components/sidebar/Sidebar.es';
 
 export default function FieldsSidebarBody({keywords, setKeywords}) {
 	const {allowFieldSets, tabs = []} = useConfig();
