+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/js/plugins/rules-sidebar/index.js
@@ -14,7 +14,7 @@
 
 import React from 'react';
 
-import {Component} from '../../../js/plugins/PluginContext.es';
+import {Component} from '../PluginContext.es';
 import RulesSidebar from './components/RulesSidebar';
 
 /**
